Move CAPIImportCommand to CommandUtils and remove it as a mixin
